+++ lib/exif_writer.dart
@@ -30,11 +30,11 @@ Future<bool> writeDateTimeToExif(
         await _noExifToolDateTimeWriter(file, dateTime, mimeTypeFromHeader)) {
       return true; //If native way was able to write exif data: exit. If not, try exifTool.
     }
-
     if (mimeTypeFromExtension != mimeTypeFromHeader &&
         mimeTypeFromHeader != 'image/tiff') {
       log(
-        "DateWriter - File has a wrong extension indicating '$mimeTypeFromExtension' but actually it is '$mimeTypeFromHeader'. Exiftool would fail, skipping.\n ${file.path}",
+        "DateWriter - File has a wrong extension indicating '$mimeTypeFromExtension' but actually it is '$mimeTypeFromHeader'.\n"
+        'ExifTool would fail on this file due to extension/content mismatch. Consider running GPTH with --fix-extensions to rename files to correct extensions.\n ${file.path}',
         level: 'error',
         forcePrint: true,
       );
@@ -96,7 +96,8 @@ Future<bool> writeGpsToExif(
 
     if (mimeTypeFromExtension != mimeTypeFromHeader) {
       log(
-        "GPSWriter - File has a wrong extension indicating '$mimeTypeFromExtension' but actually it is '$mimeTypeFromHeader'. Exiftool would fail, skipping.\n ${file.path}",
+        "GPSWriter - File has a wrong extension indicating '$mimeTypeFromExtension' but actually it is '$mimeTypeFromHeader'.\n"
+        'Exiftool would fail, skipping. You may want to run GPTH with --fix-extensions.\n ${file.path}',
         level: 'error',
         forcePrint: true,
       );